+++ src/source/wave/ComplexWaveSource.java
@@ -41,7 +41,7 @@ public class ComplexWaveSource extends WaveSource
 
     private AudioInputStream mInputStream = null;
     private int mBytesPerFrame = 0;
-    private int mFrequency = 0;
+    private long mFrequency = 0;
     
     private byte[] mBuffer;
     private int mBufferPointer = 0;
@@ -86,7 +86,7 @@ public class ComplexWaveSource extends WaveSource
 	 * Returns the frequency set for this file.  Normally returns zero, but
 	 * the value can be set with setFrequency() method.
 	 */
-    public int getFrequency()
+    public long getFrequency()
     {
 	    return mFrequency;
     }
@@ -94,7 +94,7 @@ public class ComplexWaveSource extends WaveSource
     /**
      * Changes the value returned from getFrequency() for this source.
      */
-    public void setFrequency( int frequency )
+    public void setFrequency( long frequency )
     {
     	mFrequency = frequency;
     }
